refactor: move metadata to different file
