+++ app/layout.server.tsx
@@ -1,0 +1,12 @@
+import { metadata } from './metadata';
+import ClientLayout from './layout';
+
+export { metadata };
+
+export default function RootLayout({
+    children,
+}: {
+    children: React.ReactNode;
+}) {
+    return <ClientLayout>{children}</ClientLayout>;
+}
